fix: wasm-build w/o --target web
