+++ src/lib/uesave/package.json
@@ -7,11 +7,13 @@
   "files": [
     "uesave_wasm_bg.wasm",
     "uesave_wasm.js",
+    "uesave_wasm_bg.js",
     "uesave_wasm.d.ts"
   ],
   "module": "uesave_wasm.js",
   "types": "uesave_wasm.d.ts",
   "sideEffects": [
+    "./uesave_wasm.js",
     "./snippets/*"
   ]
 }
