+++ src/lib/uesave/uesave_wasm_bg.wasm.d.ts
@@ -7,4 +7,4 @@ export function __wbindgen_malloc(a: number, b: number): number;
 export function __wbindgen_realloc(a: number, b: number, c: number, d: number): number;
 export function __wbindgen_add_to_stack_pointer(a: number): number;
 export function __wbindgen_free(a: number, b: number, c: number): void;
-export function wasm_bindgen__convert__closures__invoke2_mut__hd5288f6213887d8f(a: number, b: number, c: number, d: number): void;
+export function wasm_bindgen__convert__closures__invoke2_mut__h3c6a09e7c39b39f2(a: number, b: number, c: number, d: number): void;
